test: GithubScreen(Stateless) 테스트 코드 작성
